Refactor and move logic to helpers.ts
