+++ apps/api/src/api/services/phases/handlers/helpers.ts
@@ -1,0 +1,52 @@
+import { HORIZON_URL, StellarTokenDetails } from '@packages/shared';
+import Big from 'big.js';
+import { Horizon, Networks } from 'stellar-sdk';
+import logger from '../../../../config/logger';
+import { GLMR_FUNDING_AMOUNT_RAW, PENDULUM_EPHEMERAL_STARTING_BALANCE_UNITS } from '../../../../constants/constants';
+import { API } from '../../pendulum/apiManager';
+import { multiplyByPowerOfTen } from '../../pendulum/helpers';
+
+export const horizonServer = new Horizon.Server(HORIZON_URL);
+export const NETWORK_PASSPHRASE = Networks.PUBLIC;
+
+export async function isStellarEphemeralFunded(stellarTarget: {
+  stellarTargetAccountId: string;
+  stellarTokenDetails: StellarTokenDetails;
+}): Promise<boolean> {
+  try {
+    // We check if the Stellar target account exists and has the respective trustline.
+    const account = await horizonServer.loadAccount(stellarTarget.stellarTargetAccountId);
+
+    const trustlineExists = account.balances.some(
+      (balance) =>
+        balance.asset_type === 'credit_alphanum4' &&
+        balance.asset_code === stellarTarget.stellarTokenDetails.stellarAsset.code.string &&
+        balance.asset_issuer === stellarTarget.stellarTokenDetails.stellarAsset.issuer.stellarEncoding,
+    );
+    return trustlineExists;
+  } catch (error) {
+    if (error?.toString().includes('NotFoundError')) {
+      logger.info(
+        `SpacewalkRedeemPhaseHandler: Stellar target account ${stellarTarget.stellarTargetAccountId} does not exist.`,
+      );
+    } else {
+      // We return an error here to ensure that the phase fails and can be retried.
+      throw new Error(`SpacewalkRedeemPhaseHandler: ${error?.toString()} while checking Stellar target account.`);
+    }
+
+    return false;
+  }
+}
+
+export async function isPendulumEphemeralFunded(pendulumEphemeralAddress: string, pendulumNode: API): Promise<boolean> {
+  const fundingAmountUnits = Big(PENDULUM_EPHEMERAL_STARTING_BALANCE_UNITS);
+  const fundingAmountRaw = multiplyByPowerOfTen(fundingAmountUnits, pendulumNode.decimals).toFixed();
+  const { data: balance } = await pendulumNode.api.query.system.account(pendulumEphemeralAddress);
+
+  return Big(balance.free.toString()).gte(fundingAmountRaw);
+}
+
+export async function isMoonbeamEphemeralFunded(moonbeamEphemeralAddress: string, moonebamNode: API): Promise<boolean> {
+  const { data: balance } = await moonebamNode.api.query.system.account(moonbeamEphemeralAddress);
+  return Big(balance.free.toString()).gte(GLMR_FUNDING_AMOUNT_RAW);
+}
